Crate aob_common

Structs

DynamicNeedle

The run-time variant of a Needle.

Error

Describes errors encountered when parsing custom pattern syntax.

Find

An Iterator for finding subsequent matches of a Needle in a haystack.

Match

Represents a matching Needle found in the haystack.

StaticNeedle

The compile-time variant of a Needle.

Enums

Method

The method chosen to quickly compare strings for equality, in lieu of strcmp, since we need to account for wildcards.

Reason

A Reason gives more context about why parsing failed.

Traits

Needle

The common interface for searching haystacks with needles.